+++ internal/model/aff_campaign.go
@@ -118,7 +118,7 @@ func (m *AffCampaign) ToDto() dto.AffCampaignDto {
 
 type AffCampaignApp struct {
 	ID                uint64                 `gorm:"primarykey" json:"id"`
-	BrandId           uint64                 `json:"brand_id"`
+	BrandId           uint                   `json:"brand_id"`
 	Brand             Brand                  `json:"brand" gorm:"foreignKey:BrandId"`
 	AccessTradeId     string                 `json:"accesstrade_id" gorm:"column:accesstrade_id"`
 	CreatedAt         time.Time              `json:"created_at"`
@@ -128,7 +128,7 @@ type AffCampaignApp struct {
 	StartTime         *time.Time             `json:"start_time"`
 	EndTime           *time.Time             `json:"end_time"`
 	StellaDescription datatypes.JSON         `json:"stella_description" gorm:"serializer:json"`
-	CategoryId        uint64                 `json:"category_id"`
+	CategoryId        uint                   `json:"category_id"`
 	StellaStatus      string                 `json:"stella_status"`
 	Thumbnail         string                 `json:"thumbnail"`
 	StellaMaxCom      string                 `json:"stella_max_com"`
@@ -167,7 +167,7 @@ func (m *AffCampaignApp) ToAffCampaignAppDto() dto.AffCampaignAppDto {
 type AffCampaignComBrand struct {
 	ID           uint64 `gorm:"primarykey" json:"id"`
 	Name         string `json:"name"`
-	BrandId      uint64 `json:"brand_id"`
+	BrandId      uint   `json:"brand_id"`
 	Brand        Brand  `json:"brand" gorm:"foreignKey:BrandId"`
 	StellaMaxCom string `json:"stella_max_com"`
 }
@@ -189,7 +189,7 @@ func (m *AffCampaignComBrand) ToAffCampaignLessDto() dto.AffCampaignLessDto {
 type AffCampComFavBrand struct {
 	ID            uint64            `gorm:"primarykey" json:"id"`
 	Name          string            `json:"name"`
-	BrandId       uint64            `json:"brand_id"`
+	BrandId       uint              `json:"brand_id"`
 	FavoriteBrand UserFavoriteBrand `gorm:"foreignKey:BrandId;references:BrandId" json:"favrorite_brand"`
 	StellaMaxCom  string            `json:"stella_max_com"`
 }
